requirements.txt
pytz
yfinance
requests
python-telegram-bot==13.15